+++ api/.gitignore
@@ -13,3 +13,6 @@ venv/
 */__pycache__/*
 */*/__pycache__/*
 */*/*/__pycache__/*
+
+
+docker-data
