+++ api/docker-compose.yaml
@@ -1,0 +1,20 @@
+version: "3.7"
+services:
+  postgres:
+    image: postgres:12
+    restart: unless-stopped
+    environment:
+      POSTGRES_DB: db
+      POSTGRES_USER: postgres
+      POSTGRES_PASSWORD: pg-password
+    ports:
+      - "5432:5432"
+    volumes:
+      - ./docker-data/postgresql:/var/lib/postgresql/data
+      # copy the sql script to create tables
+      # - ./sql/create_tables.sql:/docker-entrypoint-initdb.d/create_tables.sql
+
+  adminer:
+    image: adminer
+    ports:
+      - 8090:8080
